--- app/src/main/java/woowacourse/movie/BundleKeys.kt
@@ -1,9 +1,0 @@
-package woowacourse.movie
-
-object BundleKeys {
-    const val MOVIE_DATA_KEY = "movieData"
-    const val MOVIE_BOOKING_INFO_KEY = "movieBookingInfo"
-    const val MOVIE_BOOKING_SEAT_INFO_KEY = "movieBookingSeatInfo"
-    const val SETTING_PUSH_ALARM_SWITCH_KEY = "settingPushAlarmSwitchKey"
-    const val ALARM_NOTIFICATION_ID = "notificationId"
-}
