+++ app/src/main/java/woowacourse/movie/fragment/movielist/MovieRecyclerViewAdapter.kt
@@ -1,9 +1,9 @@
-package woowacourse.movie.movielist
+package woowacourse.movie.fragment.movielist
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import woowacourse.movie.Ad
+import woowacourse.movie.model.Ad
 import woowacourse.movie.R
 import woowacourse.movie.movie.MovieUIModel
