+++ app/src/main/java/woowacourse/movie/fragment/movielist/ViewType.kt
@@ -1,0 +1,5 @@
+package woowacourse.movie.fragment.movielist
+
+enum class ViewType {
+    MOVIE, AD
+}
