--- app/src/main/java/woowacourse/movie/movielist/ViewType.kt
@@ -1,5 +1,0 @@
-package woowacourse.movie.movielist
-
-enum class ViewType {
-    MOVIE, AD
-}
