+++ app/src/main/res/layout/fragment_book_history.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragment.BookHistoryFragment">
+    tools:context=".fragment.bookhistory.BookHistoryFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerview_book_history_list"
