+++ domain/src/main/java/com/woowacourse/domain/movie/MovieBookingInfo.kt
@@ -1,4 +1,4 @@
-package com.woowacourse.domain
+package com.woowacourse.domain.movie
 
 data class MovieBookingInfo(
     val movieInfo: Movie,
