+++ domain/src/main/java/com/woowacourse/domain/movie/MovieBookingSeatInfo.kt
@@ -1,4 +1,4 @@
-package com.woowacourse.domain
+package com.woowacourse.domain.movie
 
 data class MovieBookingSeatInfo(
     val movieBookingInfo: MovieBookingInfo,
